requirements.txt
google-api-python-client
confluent-kafka
pymongo
streamlit
textblob
pyspark
numpy
pandas
plotly
scikit-learn
hdfs
python-dotenv
vaderSentiment